requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
numpy==2.1.2